refactor: Extracting interfaces from DerivationEngine class
