+++ src/libs/model/derivation/enums/edge_direction.enum.ts
@@ -1,0 +1,5 @@
+export enum EdgeDirection {
+    OUTBOUND,
+    INBOUND,
+    BIDIRECTIONAL,
+}
