+++ src/libs/model/derivation/enums/edge_scope.enum.ts
@@ -1,0 +1,5 @@
+export enum EdgeScope {
+    NON_DERIVED_ONLY,
+    DERIVED_ONLY,
+    ALL,
+}
