+++ src/libs/model/derivation/enums/rule_part.enum.ts
@@ -1,0 +1,5 @@
+export enum RulePart {
+    FIRST_PART_ELEMENT,
+    MIDDLE_ELEMENT,
+    SECOND_PART_ELEMENT,
+}
