+++ src/libs/model/derivation/rule_conditional.interface.ts
@@ -1,0 +1,7 @@
+import {RuleEdgeDescription} from "@libs/model/derivation/rule_edge_description.interface";
+
+export interface RuleConditional {
+    firstPart: RuleEdgeDescription;
+    middleElementTypes: Array<string>;
+    secondPart: RuleEdgeDescription;
+}
