+++ src/libs/model/derivation/rule_edge_description.interface.ts
@@ -1,0 +1,7 @@
+import {EdgeDirection} from "@libs/engine/derivation_engine/derivation_rule.class";
+
+export interface RuleEdgeDescription {
+    elementTypes: Array<string>;
+    edgeTypes: Array<string>;
+    direction: EdgeDirection;
+}
